requirement.txt
altgraph==0.17.2
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
ddddocr==1.4.3
et-xmlfile==1.1.0
flatbuffers==2.0
h11==0.13.0
idna==3.3
macholib==1.16
numpy==1.22.3
onnxruntime==1.11.1
opencv-python==4.5.5.64
openpyxl==3.0.9
outcome==1.1.0
pandas==1.4.2
Pillow==9.1.0
protobuf==3.20.1
pycparser==2.21
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
webdriver-manager==3.5.4
wsproto==1.1.0
xlrd==2.0.1
XlsxWriter==3.0.3
